I updated the HTML and CSS
i slightly changed the structure of the HTML so it would be easier to work on the other sites of the page (e.g. I added more classes so that I could easily copy it to other parts, using the same CSS). 
I also included a header and changed the design a bit.